[DB/sql_management] Updated teardown script:     Added dependency to get user confirmation before running teardown script

+++ server/DB/sql_management/teardown.js
@@ -1,17 +1,24 @@
 /**
  * Module to tear down everything created using setup.js
+ * Or you can simply drop the CE database from pgAdmin
  * @author JJ
  */
 
 require("dotenv").config();
 
-const config = require("./knexfile")[process.env.NODE_ENV || "development"];
-const dbName = `${config.connection.database}`; // Use string interpolation to "clone" the value
-config.connection.database = null; // change to null to connect to default DB to modify the actual targetted DB
-const knex = require("knex")(config);
-
 (async function teardown() {
+  let knex = {}; // Only assign in try/catch but needed outside of block
+
   try {
+    const yesno = require("yesno");
+    if (!(await yesno({ question: "Are you sure you want to continue?" })))
+      return;
+
+    const config = require("./knexfile")[process.env.NODE_ENV || "development"];
+    const dbName = `${config.connection.database}`; // Use string interpolation to "clone" the value
+    config.connection.database = null; // change to null to connect to default DB to modify the actual targetted DB
+    knex = require("knex")(config);
+
     // Drop the database only if it exists
     // DB name is case sensitive thus in double quotes
     await knex.raw(`DROP DATABASE IF EXISTS "${dbName}";`);
